refactor: move migrations logic to utils
